feat(eslint): arm a table-teardown sweep on the dropTable() helper form
require-table-teardown recognised the drop half of a catalogue sweep only
in raw SQL, so a file that sweeps through the schema-statement helper —
the form CLAUDE.md steers tests toward — reported every prefixed create it
does tear down.

Extract require-canonical-rebuild's sweep-binding resolver into
eslint/sweep-binding.mjs and use it from both rules, so a dropTable() call
whose argument traces back to a sink-derived row binding arms a sweep on
the same footing as the raw spelling. A fixed-name dropTable("ex_foo")
still arms nothing.
